DDP-5609 | moved things to Dao class, added check if realm has ability of adding family member
