+++ src/main/java/org/broadinstitute/dsm/db/DDPInstance.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.NonNull;
 import org.apache.commons.lang3.StringUtils;
 import org.broadinstitute.ddp.db.SimpleResult;
+import org.broadinstitute.dsm.db.dao.ddp.instance.DDPInstanceDao;
 import org.broadinstitute.dsm.model.Value;
 import org.broadinstitute.dsm.statics.DBConstants;
 import org.broadinstitute.dsm.statics.QueryExtension;
@@ -25,11 +26,6 @@ public class DDPInstance {
 
     private static final Logger logger = LoggerFactory.getLogger(DDPInstance.class);
 
-    private static final String SQL_SELECT_INSTANCE_WITH_ROLE = "SELECT ddp_instance_id, instance_name, base_url, collaborator_id_prefix, migrated_ddp, billing_reference, " +
-            "es_participant_index, es_activity_definition_index, es_users_index, (SELECT count(role.name) " +
-            "FROM ddp_instance realm, ddp_instance_role inRol, instance_role role WHERE realm.ddp_instance_id = inRol.ddp_instance_id AND inRol.instance_role_id = role.instance_role_id AND role.name = ? " +
-            "AND realm.ddp_instance_id = main.ddp_instance_id) AS 'has_role', mr_attention_flag_d, tissue_attention_flag_d, auth0_token, notification_recipients FROM ddp_instance main " +
-            "WHERE is_active = 1";
     private static final String SQL_SELECT_INSTANCE_WITH_KIT_BEHAVIOR = "SELECT main.ddp_instance_id, instance_name, base_url, collaborator_id_prefix, migrated_ddp, billing_reference, " +
             "es_participant_index, es_activity_definition_index, es_users_index, mr_attention_flag_d, tissue_attention_flag_d, auth0_token, notification_recipients, kit_behavior_change " +
             "FROM ddp_instance main, instance_settings setting WHERE main.ddp_instance_id = setting.ddp_instance_id " +
@@ -165,7 +161,7 @@ public class DDPInstance {
     public static DDPInstance getDDPInstanceWithRole(@NonNull String realm, @NonNull String role) {
         SimpleResult results = inTransaction((conn) -> {
             SimpleResult dbVals = new SimpleResult();
-            try (PreparedStatement stmt = conn.prepareStatement(SQL_SELECT_INSTANCE_WITH_ROLE + QueryExtension.BY_INSTANCE_NAME)) {
+            try (PreparedStatement stmt = conn.prepareStatement(DDPInstanceDao.SQL_SELECT_INSTANCE_WITH_ROLE + QueryExtension.BY_INSTANCE_NAME)) {
                 stmt.setString(1, role);
                 stmt.setString(2, realm);
                 try (ResultSet rs = stmt.executeQuery()) {
@@ -219,7 +215,7 @@ public class DDPInstance {
         List<DDPInstance> ddpInstances = new ArrayList<>();
         SimpleResult results = inTransaction((conn) -> {
             SimpleResult dbVals = new SimpleResult();
-            try (PreparedStatement bspStatement = conn.prepareStatement(SQL_SELECT_INSTANCE_WITH_ROLE)) {
+            try (PreparedStatement bspStatement = conn.prepareStatement(DDPInstanceDao.SQL_SELECT_INSTANCE_WITH_ROLE)) {
                 bspStatement.setString(1, role);
                 try (ResultSet rs = bspStatement.executeQuery()) {
                     while (rs.next()) {
@@ -238,33 +234,6 @@ public class DDPInstance {
             throw new RuntimeException("Error looking ddpInstances ", results.resultException);
         }
         return ddpInstances;
-    }
-
-    public static boolean getRole(@NonNull String realm, @NonNull String role) {
-        SimpleResult results = inTransaction((conn) -> {
-            SimpleResult dbVals = new SimpleResult();
-            try (PreparedStatement stmt = conn.prepareStatement(SQL_SELECT_INSTANCE_WITH_ROLE + QueryExtension.BY_INSTANCE_NAME)) {
-                stmt.setString(1, role);
-                stmt.setString(2, realm);
-                try (ResultSet rs = stmt.executeQuery()) {
-                    if (rs.next()) {
-                        dbVals.resultValue = rs.getBoolean(DBConstants.HAS_ROLE);
-                    }
-                }
-                catch (SQLException e) {
-                    throw new RuntimeException("Error getting role of realm " + realm, e);
-                }
-            }
-            catch (SQLException ex) {
-                dbVals.resultException = ex;
-            }
-            return dbVals;
-        });
-
-        if (results.resultException != null) {
-            throw new RuntimeException("Couldn't get role of realm " + realm, results.resultException);
-        }
-        return (boolean) results.resultValue;
     }
 
     private static DDPInstance getDDPInstanceWithRoleFormResultSet(@NonNull ResultSet rs) throws SQLException {
